dedupe data fetching on dashboard
