+++ package.json
@@ -35,9 +35,9 @@
 		"valibot": "^0.30.0"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "^1.7.1",
+		"@biomejs/biome": "^1.7.2",
 		"@next/bundle-analyzer": "^14.2.3",
-		"@types/node": "^20.12.7",
+		"@types/node": "^20.12.8",
 		"@types/react": "^18.3.1",
 		"@types/react-dom": "^18.3.0",
 		"@types/textarea-caret": "^3.0.3",
@@ -49,7 +49,7 @@
 		"tailwindcss": "^3.4.3",
 		"typescript": "~5.4.5"
 	},
-	"packageManager": "pnpm@8.15.7+sha256.50783dd0fa303852de2dd1557cd4b9f07cb5b018154a6e76d0f40635d6cee019",
+	"packageManager": "pnpm@8.15.8+sha256.691fe176eea9a8a80df20e4976f3dfb44a04841ceb885638fe2a26174f81e65e",
 	"engines": {
 		"node": "20"
 	}
