+++ src/app/guilds/[guildId]/emojis/page.tsx
@@ -1,10 +1,9 @@
+import { getGuildSettings } from "@/app/guilds/[guildId]/get-guild-data.ts";
 import { Form } from "@/components/dashboard/Form.tsx";
 import { Section } from "@/components/dashboard/Section.tsx";
 import { Text } from "@/components/dashboard/Text.tsx";
-import type { Guild, GuildSettings } from "@/lib/guild.ts";
 import { TOKEN_COOKIE } from "@/utils/constants.ts";
 import type { Snowflake } from "@/utils/discord-cdn.ts";
-import { makeApiRequest } from "@/utils/make-api-request.ts";
 import { cookies } from "next/headers";
 import { EmojiListChannel } from "./01-emoji-list-channel.tsx";
 import { UpdateEmojiList } from "./02-update-emoji-list.tsx";
@@ -12,7 +11,7 @@ import { update } from "./update.ts";
 
 export default async function Emojis({ params: { guildId } }: { readonly params: { guildId: Snowflake } }) {
 	const token = cookies().get(TOKEN_COOKIE)!.value;
-	const { guild, settings } = await getData(guildId, token);
+	const { guild, settings } = await getGuildSettings(guildId, token, "emojis");
 
 	const action = update.bind(null, guildId);
 
@@ -38,23 +37,4 @@ export default async function Emojis({ params: { guildId } }: { readonly params:
 			</Section>
 		</Form>
 	);
-}
-
-async function getData(guildId: Snowflake, token: string) {
-	const response = await makeApiRequest(`/guilds/${guildId}`, token, {
-		next: {
-			tags: [`settings:${guildId}`, `settings:${guildId}:emojis`],
-		},
-	});
-
-	if (!response.ok) {
-		throw new Error("Failed to get guild information");
-	}
-
-	return response.json() as Promise<GetGuildResponse>;
-}
-
-interface GetGuildResponse {
-	guild: Guild;
-	settings: GuildSettings;
 }
