+++ src/app/guilds/[guildId]/milestones/page.tsx
@@ -1,11 +1,10 @@
+import { getGuildSettings } from "@/app/guilds/[guildId]/get-guild-data.ts";
 import { Separator } from "@/components/Separator.tsx";
 import { Form } from "@/components/dashboard/Form.tsx";
 import { Section } from "@/components/dashboard/Section.tsx";
 import { Text } from "@/components/dashboard/Text.tsx";
-import type { Guild, GuildSettings } from "@/lib/guild.ts";
 import { TOKEN_COOKIE } from "@/utils/constants.ts";
 import type { Snowflake } from "@/utils/discord-cdn.ts";
-import { makeApiRequest } from "@/utils/make-api-request.ts";
 import { cookies } from "next/headers";
 import { MilestonesChannel } from "./01-milestones-channel.tsx";
 import { MilestoneInterval } from "./02-milestone-interval.tsx";
@@ -15,7 +14,7 @@ import { update } from "./update.ts";
 
 export default async function Milestones({ params: { guildId } }: { readonly params: { guildId: Snowflake } }) {
 	const token = cookies().get(TOKEN_COOKIE)!.value;
-	const { guild, settings } = await getData(guildId, token);
+	const { guild, settings } = await getGuildSettings(guildId, token, "milestones");
 
 	const action = update.bind(null, guildId, settings.premium);
 
@@ -67,23 +66,4 @@ export default async function Milestones({ params: { guildId } }: { readonly par
 			</Section>
 		</Form>
 	);
-}
-
-async function getData(guildId: Snowflake, token: string) {
-	const response = await makeApiRequest(`/guilds/${guildId}`, token, {
-		next: {
-			tags: [`settings:${guildId}`, `settings:${guildId}:milestones`],
-		},
-	});
-
-	if (!response.ok) {
-		throw new Error("Failed to get guild information");
-	}
-
-	return response.json() as Promise<GetGuildResponse>;
-}
-
-interface GetGuildResponse {
-	guild: Guild;
-	settings: GuildSettings;
 }
