+++ src/app/guilds/[guildId]/miscellaneous/page.tsx
@@ -1,11 +1,10 @@
+import { getGuildSettings } from "@/app/guilds/[guildId]/get-guild-data.ts";
 import { Separator } from "@/components/Separator.tsx";
 import { Form } from "@/components/dashboard/Form.tsx";
 import { Section } from "@/components/dashboard/Section.tsx";
 import { Text } from "@/components/dashboard/Text.tsx";
-import type { Guild, GuildSettings } from "@/lib/guild.ts";
 import { TOKEN_COOKIE } from "@/utils/constants.ts";
 import type { Snowflake } from "@/utils/discord-cdn.ts";
-import { makeApiRequest } from "@/utils/make-api-request.ts";
 import { cookies } from "next/headers";
 import { StoreMemberCountData } from "./01-store-member-count-data.tsx";
 import { AutomaticallyPublishChannels } from "./02-automatically-publish-channels.tsx";
@@ -13,7 +12,7 @@ import { update } from "./update.ts";
 
 export default async function Miscellaneous({ params: { guildId } }: { readonly params: { guildId: Snowflake } }) {
 	const token = cookies().get(TOKEN_COOKIE)!.value;
-	const { guild, settings } = await getData(guildId, token);
+	const { guild, settings } = await getGuildSettings(guildId, token, "miscellaneous");
 
 	const action = update.bind(null, guildId, settings.premium);
 
@@ -43,23 +42,4 @@ export default async function Miscellaneous({ params: { guildId } }: { readonly 
 			</Section>
 		</Form>
 	);
-}
-
-async function getData(guildId: Snowflake, token: string) {
-	const response = await makeApiRequest(`/guilds/${guildId}`, token, {
-		next: {
-			tags: [`settings:${guildId}`, `settings:${guildId}:miscellaneous`],
-		},
-	});
-
-	if (!response.ok) {
-		throw new Error("Failed to get guild information");
-	}
-
-	return response.json() as Promise<GetGuildResponse>;
-}
-
-interface GetGuildResponse {
-	guild: Guild;
-	settings: GuildSettings;
 }
